Create configuration only for running VMs
The `vagrant ssh-config` will fail of one of the VMs is not running. To
avoid having to keep all of the VMs running, the network configuration
script can query the set of running VMs first and only generate the
configuration for those.

Also fixed the hostnames for the node and galera type VMS to be
valid. Previously it used underscores which wasn't what was actually
used on the VM.