Use unsafe caching for libvirt disks
Given that there's no need for fsyncs to actually fsync the data, using
'unsafe' speeds things up. Doesn't really make a difference on a fast
SSDs but on slower storage it makes a huge difference.

Also increased the base memory to 1GiB, this avoids the case where the
MariaDB servers are constantly swapping which makes the problems worse
on slow storage.