Add scripts for setups using Incus
System containers are an extremely light approach that provide a near
identical experience to full virtualization via Vagrant. The benefits of
this are similar to application containers (i.e. docker) where only as
much memory is used as is needed.

This approach enables dense packing of test setups and makes parallel
testing possible.